Re-create your version of W3
For this recreation I used my background as a pastry
chef to teach others how to make sourdough bread. The
index.html file is a general introduction to the bread
I am attempting to give intructions to. I list some
common termonology and include links to other sites I
found useful as well as the link to my what-is.html file
that explains how to make a key component for the bread
itself. Lastly, on the what-is.html, I include a link to
my overview.html where I give a recipe for a beginner
sourdough and list the steps to making a successful loaf.

I hope this finds everyone well and hope you try it out!